% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.R
\name{custom}
\alias{custom}
\title{Custom map profile}
\usage{
custom(FUN = mean, attribute, units = "")
}
\arguments{
\item{FUN}{A summary function to be called on the provided event attribute, e.g. mean, median, min, max}

\item{attribute}{The name of the case attribute to visualize}
}
\description{
Function to create a custom map profile based on some event log attribute.
}
\examples{
\dontrun{
library(eventdataR)
library(processmapR)
data(traffic_fines)
# make sure the amount attribute is propagated forward in each trace
# note that zoo::na.locf is much faster than tidyr::fill still it is incredibly slow for what it does
library(zoo)
traffic_fines_prepared <- traffic_fines \%>\% group_by_case() \%>\% mutate(amount = na.locf(amount, na.rm = F)) \%>\% ungroup_eventlog() \%>\% filter_trace_frequency(percentage = 0.8)
process_map(traffic_fines_prepared, type_nodes = custom(attribute = "amount", units = "EUR"))
}

}
