% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/dotted_char_plotly_i.R
\name{deprecated}
\alias{deprecated}
\alias{idotted_chart}
\alias{iplotly_dotted_chart}
\title{Deprecated Functions}
\usage{
idotted_chart(eventlog, plotly = FALSE)

iplotly_dotted_chart(eventlog)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
