% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_graph.R
\docType{import}
\name{export_graph}
\alias{export_graph}
\title{Export a graph to various image formats}
\arguments{
\item{graph}{A graph object of class `dgr_graph`.}

\item{file_name}{The name of the exported file (including it's extension).}

\item{file_type}{The type of file to be exported. Options for graph files
are: `png`, `pdf`, `svg`, and `ps`.}

\item{title}{An optional title for the output graph.}

\item{width}{Output width in pixels or `NULL` for default. Only useful for
export to image file formats `png`, `pdf`, `svg`, and `ps`.}

\item{height}{Output height in pixels or `NULL` for default. Only useful for
export to image file formats `png`, `pdf`, `svg`, and `ps`.}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DiagrammeR}{\code{\link[DiagrammeR]{export_graph}}}
}}

