% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_map.R
\name{export_map}
\alias{export_map}
\title{Export process map to pdf, png, ps or svg.}
\usage{
export_map(
  map,
  file_name = NULL,
  file_type = NULL,
  title = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{map}{A \code{process_map} created with \code{\link[=process_map]{process_map}} and argument \code{render = F}.}

\item{file_name}{The name of the exported file (including it's extension).}

\item{file_type}{The type of file to be exported. Options for graph files
are: \code{png}, \code{pdf}, \code{svg}, and \code{ps}.}

\item{title}{An optional title for the output graph.}

\item{width}{Output width in pixels or \code{NULL} for default. Only useful for
export to image file formats \code{png}, \code{pdf}, \code{svg}, and \code{ps}.}

\item{height}{Output height in pixels or \code{NULL} for default. Only useful for
export to image file formats \code{png}, \code{pdf}, \code{svg}, and \code{ps}.}
}
\description{
Export process map to pdf, png, ps or svg.
}
