% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meta_data.R
\name{get_activities}
\alias{get_activities}
\alias{get_flows}
\title{Get values for nodes and flows from process map}
\usage{
get_activities(process_map)

get_flows(process_map)
}
\arguments{
\item{process_map}{An object created using process_map function. Can both be a rendered or not rendered object.}
}
\description{
Get values for nodes and flows from process map
}
