% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\title{Performance map profile}
\usage{
performance(FUN = mean, units = "days", flow_time = c("inter_start_time",
  "idle_time"))
}
\arguments{
\item{FUN}{A summary function to be called on the process time of a specific activity, e.g. mean, median, min, max}

\item{units}{The time unit in which processing time should be presented (mins, hours, days, weeks)}
}
\description{
Function to create a performance map profile to be used as the type of a process map. It results in a process map describing process time.
}
