% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precedence_matrix.plot.R
\name{plot.process_matrix}
\alias{plot.process_matrix}
\title{Process Matrix Plot}
\usage{
\method{plot}{process_matrix}(x, ...)
}
\arguments{
\item{x}{Precedence matrix}

\item{...}{Additional paramters}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize a precendence matrix. A generic plot function for precedences matrices.
}
