% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMapOutput.R
\name{processMapOutput}
\alias{processMapOutput}
\title{Widget output function for use in Shiny}
\usage{
processMapOutput(outputId, width = "100\%", heigth = "400px")
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width}{A valid CSS unit for the width or a number,
which will be coerced to a string and have px appended.}

\item{heigth}{A valid CSS unit for the height or a number,
which will be coerced to a string and have px appended.}
}
\description{
Widget output function for use in Shiny
}
