% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processmapR.R
\docType{package}
\name{processmapR}
\alias{processmapR}
\alias{processmapR-package}
\title{processmapR - Process Maps in R}
\description{
This package provides several useful techniques process visualization.
}
