% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_map.R
\name{render_map}
\alias{render_map}
\title{Render process map}
\usage{
render_map(
  map,
  layout = NULL,
  output = NULL,
  as_svg = FALSE,
  title = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{map}{A \code{process_map} created with \code{\link[=process_map]{process_map}} and argument \code{render = F}.}

\item{layout}{A string specifying a layout type to use for node placement in
this rendering. Possible layouts include: \code{nicely}, \code{circle}, \code{tree}, \code{kk},
and \code{fr}.}

\item{output}{A string specifying the output type; \code{graph} (the default)
renders the graph using the \code{\link[DiagrammeR:grViz]{grViz()}} function and \code{visNetwork} renders the
graph using the \code{\link[DiagrammeR:visnetwork]{visnetwork()}} function.}

\item{as_svg}{An option to render the graph as an SVG document.}

\item{title}{An optional title for a graph when using \code{output = "graph"}.}

\item{width}{An optional parameter for specifying the width of the resulting
graphic in pixels.}

\item{height}{An optional parameter for specifying the height of the
resulting graphic in pixels.}
}
\description{
Render process map
}
