% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_explorer.R
\name{trace_explorer}
\alias{trace_explorer}
\alias{trace_explorer.eventlog}
\alias{trace_explorer.activitylog}
\alias{plotly_trace_explorer}
\title{Trace Explorer}
\usage{
trace_explorer(
  log,
  coverage = NULL,
  n_traces = NULL,
  type = c("frequent", "infrequent"),
  coverage_labels = c("relative", "absolute", "cumulative"),
  abbreviate = TRUE,
  show_labels = TRUE,
  label_size = 3,
  scale_fill = bupaR::scale_fill_discrete_bupaR(),
  raw_data = FALSE,
  eventlog = deprecated(),
  .abbreviate = deprecated()
)

\method{trace_explorer}{eventlog}(
  log,
  coverage = NULL,
  n_traces = NULL,
  type = c("frequent", "infrequent"),
  coverage_labels = c("relative", "absolute", "cumulative"),
  abbreviate = TRUE,
  show_labels = TRUE,
  label_size = 3,
  scale_fill = bupaR::scale_fill_discrete_bupaR(),
  raw_data = FALSE,
  eventlog = deprecated(),
  .abbreviate = deprecated()
)

\method{trace_explorer}{activitylog}(
  log,
  coverage = NULL,
  n_traces = NULL,
  type = c("frequent", "infrequent"),
  coverage_labels = c("relative", "absolute", "cumulative"),
  abbreviate = TRUE,
  show_labels = TRUE,
  label_size = 3,
  scale_fill = bupaR::scale_fill_discrete_bupaR(),
  raw_data = FALSE,
  eventlog = deprecated(),
  .abbreviate = deprecated()
)

plotly_trace_explorer(
  log,
  coverage = NULL,
  n_traces = NULL,
  type = c("frequent", "infrequent"),
  coverage_labels = c("relative", "absolute", "cumulative"),
  abbreviate = TRUE,
  show_labels = TRUE,
  label_size = 3,
  scale_fill = bupaR::scale_fill_discrete_bupaR(),
  raw_data = FALSE,
  eventlog = deprecated(),
  .abbreviate = deprecated()
)
}
\arguments{
\item{log}{\code{\link[bupaR]{log}}: Object of class \code{\link[bupaR]{log}} or derivatives (\code{\link[bupaR]{grouped_log}}, \code{\link[bupaR]{eventlog}}, \code{\link[bupaR]{activitylog}}, etc.).}

\item{coverage}{\code{\link{numeric}} (default \code{0.2}): The percentage coverage of the trace to explore. Defaults to \code{0.2} (\code{0.05}) most (in)frequent.}

\item{n_traces}{\code{\link{integer}}: Instead of setting \code{coverage}, an exact number of traces can be set. Should be an \code{\link{integer}} larger than \code{0}.}

\item{type}{\code{\link{character}} (default \code{"frequent"}): \code{"frequent"} traces first, or \code{"infrequent"} traces first?}

\item{coverage_labels}{\code{\link{character}} (default \code{"relative"}): Change the labels to be shown on the right of the process variants.
These can be \code{"relative"} frequency (default), \code{"absolute"}, or \code{"cumulative"}.}

\item{abbreviate}{\code{\link{logical}} (default \code{TRUE}): If \code{TRUE}, abbreviate activity labels.}

\item{show_labels}{\code{\link{logical}} (default \code{TRUE}): If \code{FALSE}, activity labels are not shown.}

\item{label_size}{\code{\link{numeric}} (default \code{3}): Font size of labels.}

\item{scale_fill}{Set color scale. Defaults to \code{\link[bupaR:scale_fill_discrete_bupaR]{scale_fill_discrete_bupaR}}.}

\item{raw_data}{\code{\link{logical}} (default \code{FALSE}): Return raw data.}

\item{eventlog}{Deprecated. Use log instead.}

\item{.abbreviate}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{abbreviate} instead.}
}
\description{
Explore traces, ordered by relative trace frequency.
}
\section{Methods (by class)}{
\itemize{
\item \code{trace_explorer(eventlog)}: Trace explorer for an \code{\link[bupaR:eventlog]{eventlog}}.

\item \code{trace_explorer(activitylog)}: Trace explorer for an \code{\link[bupaR:activitylog]{activitylog}}.

}}
