% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/dotted_char_plotly_i.R,
%   R/lined_chart_plotly_i.R, R/trace_explorer.R
\name{deprecated}
\alias{deprecated}
\alias{idotted_chart}
\alias{iplotly_dotted_chart}
\alias{plotly_dotted_chart}
\alias{ilined_chart}
\alias{iplotly_lined_chart}
\alias{plotly_lined_chart}
\alias{plotly_trace_explorer}
\title{Deprecated Functions}
\usage{
idotted_chart(eventlog, plotly = FALSE)

iplotly_dotted_chart(eventlog)

plotly_dotted_chart(
  log,
  x = c("absolute", "relative", "relative_week", "relative_day"),
  sort = c("auto", "start", "end", "duration", "start_week", "start_day"),
  color = NULL,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  add_end_events = FALSE,
  scale_color = bupaR::scale_color_discrete_bupaR,
  eventlog = deprecated()
)

ilined_chart(eventlog, plotly = FALSE)

iplotly_lined_chart(eventlog)

plotly_lined_chart(
  log,
  x = c("absolute", "relative"),
  sort = c("auto", "start", "end", "duration"),
  color = NULL,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  line_width = 2,
  plotly = TRUE,
  scale_color = bupaR::scale_color_discrete_bupaR,
  eventlog = deprecated()
)

plotly_trace_explorer(
  log,
  coverage = NULL,
  n_traces = NULL,
  type = c("frequent", "infrequent"),
  coverage_labels = c("relative", "absolute", "cumulative"),
  abbreviate = TRUE,
  show_labels = TRUE,
  label_size = 3,
  scale_fill = bupaR::scale_fill_discrete_bupaR,
  raw_data = FALSE,
  plotly = TRUE,
  eventlog = deprecated(),
  .abbreviate = deprecated()
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{\link[=plotly_dotted_chart]{plotly_dotted_chart()}} has been deprecated in favour of \code{plotly_dotted_chart(..., plotly = TRUE)}.

\code{\link[=plotly_lined_chart]{plotly_lined_chart()}} has been deprecated in favour of \code{plotly_lined_chart(..., plotly = TRUE)}.

\code{\link[=plotly_trace_explorer]{plotly_trace_explorer()}} has been deprecated in favour of \code{plotly_trace_explorer(..., plotly = TRUE)}.
}
\keyword{internal}
