% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{frequency}
\alias{frequency}
\title{Frequency map profile}
\usage{
frequency(value = c("absolute", "relative"))
}
\arguments{
\item{value}{The type of frequency value to be used: absolute or relative.}
}
\description{
Function to create a frequency profile for a process map.
}
