% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{frequency}
\alias{frequency}
\title{Frequency map profile}
\usage{
frequency(value = c("absolute", "relative", "absolute_case", "relative_case"),
  color_scale = "PuBu")
}
\arguments{
\item{value}{The type of frequency value to be used:
absolute, relative (percentage of activity instances) or relative_case (percentage of cases the activity occurs in).}

\item{color_scale}{Name of color scale to be used for nodes. Defaults to PuBu. See `Rcolorbrewer::brewer.pal.info()` for all options.}
}
\description{
Function to create a frequency profile for a process map.
}
