% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lined_chart.R
\name{lined_chart}
\alias{lined_chart}
\alias{ilined_chart}
\alias{iplotly_lined_chart}
\alias{plotly_lined_chart}
\title{Lined chart}
\usage{
lined_chart(
  eventlog,
  x = c("absolute", "relative"),
  sort = NULL,
  color,
  units,
  line_width = 2,
  plotly,
  ...
)

ilined_chart(eventlog, plotly = FALSE)

iplotly_lined_chart(eventlog)

plotly_lined_chart(eventlog, color = NULL, ...)
}
\arguments{
\item{eventlog}{Eventlog object}

\item{color}{Optional. Should be a case attribute! No coloring applied by default.}

\item{plotly}{Return plotly object}

\item{...}{Deprecated arguments}
}
\description{
Create a lined chart to view all cases at a glance
}
