% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lined_chart.R
\name{lined_chart}
\alias{lined_chart}
\alias{ilined_chart}
\alias{iplotly_lined_chart}
\alias{plotly_lined_chart}
\title{Lined chart}
\usage{
lined_chart(
  eventlog,
  x = c("absolute", "relative"),
  sort = NULL,
  color,
  units,
  line_width = 2,
  plotly,
  ...
)

ilined_chart(eventlog, plotly = FALSE)

iplotly_lined_chart(eventlog)

plotly_lined_chart(eventlog, color = NULL, ...)
}
\arguments{
\item{eventlog}{Eventlog object}

\item{x}{Value for plot on x-axis: absolute time or relative time (since start, since start of week, since start of day)}

\item{sort}{Ordering of the cases on y-axis: start, end or duration, start_week, start_day}

\item{color}{Optional. Should be a case attribute! No coloring applied by default.}

\item{units}{Time units to use on x-axis in case of relative time.}

\item{line_width}{The width of lines}

\item{plotly}{Return plotly object}

\item{...}{Deprecated arguments}
}
\description{
Create a lined chart to view all cases at a glance
}
