% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_matrix.R
\name{process_matrix}
\alias{process_matrix}
\alias{process_matrix.eventlog}
\title{Create process matrix}
\usage{
process_matrix(eventlog, type, ...)

\method{process_matrix}{eventlog}(eventlog, type = frequency(), ...)
}
\arguments{
\item{eventlog}{The event log object for which to create a process matrix}

\item{type}{A process matrix type, which can be created with the functions frequency, performance and custom. The first type focusses on the frequency aspect of a process, while the second one focussed on processing time. The third one allows custom attributes to be used.}

\item{...}{Other arguments}
}
\description{
Create process matrix
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Process matrix for event log
}}

