% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_map.R
\name{resource_map}
\alias{resource_map}
\title{Resource Map}
\usage{
resource_map(eventlog, render = T)
}
\arguments{
\item{eventlog}{The event log object for which to create a resource map}

\item{render}{Whether the map should be rendered immediately (default), or rather an object of type dgr_graph should be returned.}
}
\description{
A function for creating a resource map of an event log based on handover of work.
}
\examples{
\dontrun{
library(eventdataR)
data(patients)
resource_map(patients)
}
}
